Moved create review query into data layer class
