+++ backend/src/review/review.module.ts
@@ -4,11 +4,12 @@ import { DatabaseService } from 'src/database/database.service';
 import { ReviewController } from './review.controller';
 import { ReviewService } from './review.service';
 import { ReviewSimilarityService } from './review-text-match';
+import {ReviewModel} from "./models/review-data-layer";
 
 @Module({
   imports: [CaptchaModule],
   controllers: [ReviewController],
-  providers: [ReviewService, DatabaseService, ReviewSimilarityService],
+  providers: [ReviewService, DatabaseService, ReviewSimilarityService, ReviewModel],
   exports: [ReviewService],
 })
 export class ReviewModule {}
